scripts/requirements.txt
torch>=1.0
torchvision
tensorboardX
pillow